adding some additional error handling

+++ DistributedTransactions.Domain/Allocations/Trade.cs
@@ -12,12 +12,12 @@ public class Trade : AllocationsDomainModel
     public TradeDestination Destination { get; set; }
     public TradeType Type { get; set; }
     public TradeRestriction Restriction { get; set; }
-    public int? Quantity { get; set; }
+    public long? Quantity { get; set; }
     public double? Price { get; set; }
     public double? Amount { get; set; }
-    public int? CancelledQuantity { get; set; }
+    public long? CancelledQuantity { get; set; }
     public TradeDestination? NewDestination { get; set; }
     public TradeType? NewType { get; set; }
     public TradeRestriction? NewRestriction { get; set; }
-    public int? NewQuantity { get; set; }
+    public long? NewQuantity { get; set; }
 }
