+++ DistributedTransactions.Domain/Orders/BlockOrder.cs
@@ -3,6 +3,6 @@ namespace DistributedTransactions.Domain.Orders;
 public class BlockOrder : BaseOrderModel
 {
     public required string Code { get; set; }
-    public int Allocated { get; set; }
+    public long Allocated { get; set; }
     public int Accounts { get; set; }
 }
