+++ DistributedTransactions.Domain/Trades/ReplacedTrade.cs
@@ -5,5 +5,5 @@ public class ReplacedTrade : BaseTradeModel
     public TradeDestination? NewDestination { get; set; }
     public TradeType? NewType { get; set; }
     public TradeRestriction? NewRestriction { get; set; }
-    public int? NewQuantity { get; set; }
+    public long? NewQuantity { get; set; }
 }
